hypervisor: memguard: kick-out a regulated CPU from a nested IRQ before suspending
If a CPU is in the regulation loop and we hit with a suspension request,
we must first kick out the CPU from the loop to correctly restore the
EL state. Afterwards we can suspend it.

TODO: this is stil hacky, the CPU should suspend / reset itself while
returning from IRQ...

Signed-off-by: Andrea Bastoni <andrea.bastoni@tum.de>
Signed-off-by: Andrea Bastoni <abastoni@minervasys.tech>