hypervisor, arm64: main QoS-400 functionalities for ZCU102 and S32V
Add QoS general parameters and the main functionalities to set the QoS
regulation on the hardware.

Code consolidated and ported from the original work by:
- Renato Mancuso (BU) <rmancuso@bu.edu>
- Rohan Tabish <rtabish@illinois.edu>
at Boston University and University of Illinois Urbana-Champaign.

Notes:
1) Support for FPD_GPV QoS regulators is only available for the ZCU102.  On
the ZCU102, the FPD_GPV registers are accessible only from EL3. In order
for JH to be able to access these registers, the ATF should be patched
to allow read/write operations through the two services
ZCU102_QOS_READ_SMC and ZCU102_QOS_WRITE_SMC.

2) The configuration from S32V for accessing the QoS registers is what can
be considered a "standard" one. The ZCU102 is special since it was not
possible to access the QoS range from EL2 in a straightforward way (and
therefore a SMC into EL3 is required).

The following patch implements the integration of the needed SMC in ATF
to support the ZCU102: 0001-SVC-for-QoS-register-r-w-on-ZCU102.patch

Lacking a proper "bsp" layer, if more boards requires a special
treatment, the quick way of implementing a board-dependend support is
to rely on conditionally compiled code and CONFIG_MACH_<boards>.

Signed-off-by: Rohan Tabish <rtabish@illinois.edu>
Signed-off-by: Renato Mancuso (BU) <rmancuso@bu.edu>
Signed-off-by: Andrea Bastoni <andrea.bastoni@tum.de>
Signed-off-by: Andrea Bastoni <abastoni@minervasys.tech>