00100#589: [hypervisor, arm64, memguard] add support to PPIs irqs for PMU
The current version of Jailhouse was supposed to work only with PMU
irqs which were classified as SGIs (persistent across CPU resets),
this is not the case with the newer boards which support GICv3 which
has PMU irqs classified as PPIs.

The solution is straightforward: just call the function pmu_cpu_reset
inside memguard_cpu_reset which re-enables PMU interrupts.

Signed-off-by: Filippo Fontana <filippo.fontana@minervasys.tech>
Signed-off-by: Andrea Bastoni <abastoni@minervasys.tech>