00100#591: [tools, membomb] fix the condition for memguard command
`mem` is the memory mapping of the communication region inside the
memory, but it was used erroneously to check if memguard has to be
enabled.

This commit changes the if conditions which used `mem` with the `memory`
variable which is the one used when the membomb tool has to enable
memguard.

Signed-off-by: Filippo Fontana <filippo.fontana@minervasys.tech>
Signed-off-by: Andrea Bastoni <abastoni@minervasys.tech>