00100#640: [hypervisor, arm64, control] apply dirty hack for RK3588 to avoid FPEXC register
The FPXEC register does not exist on the A76 cores causing errors while
Jailhouse is executing cells on those cores. Proper handling of this
exception will be added in the future, in the mean time, this hack is
applied. The hack requires to define the CONFIG_MACH_RK3588 macro in
`include/jailhouse/config.h`.

Signed-off-by: Filippo Fontana <filippo.fontana@minervasys.tech>
Signed-off-by: Andrea Bastoni <abastoni@minervasys.tech>